refactor(server): extract JSON/LLM parsers into src/server/llm-json.js
Pure move, no behavior change. extractJSON + safeParseLLM (safeParseLLM calls
extractJSON, so they move together) relocate to src/server/llm-json.js;
server.js imports them. All call sites unchanged (9 extractJSON + 11
safeParseLLM remain in server.js; the def + the internal call moved into the
module). These are the core LLM-output parsers used by nearly every agent call.

Block moved by line-marker via script (90 lines) to avoid transcription error.

Adds test/llm-json.test.js: balanced extraction, array mode, truncation
recovery, code-fence stripping, trailing-comma recovery, total-failure throw.
The unit test is also what verifies the module loads in CI, since the gate
doesn't boot the app.

Route inventory unchanged (213). node --check clean on both files.
